add uniform size for header and remove icon home link
Used flex-boxes and size-hinting to normalize the size of all headers.

Removed the button from SBAR and Emergency note headers and replaced with an image, preventing going to the main screen without saving by clicking the icon.

Changed some spacing and padding values in pin-screen and set-screen as well as wrapping all pin rows in a wrapper to facilitate normalization of header size.